Fixed generated problems
There will be problems if the modified classes are the ones in the
src-gen source folder because when the mw2e re-generate the artifacts,
all of the edited classes will reset. So the thing is to create a class
to extend the class of the src-gen package and edit, then override the
configure bind to point the created class
